Attempting to decouple field_list from type_deduction
