+++ include/field_list/field_list.hpp
@@ -9,6 +9,7 @@
 #include "../lib/containers/fixed_string.hpp"
 #include "field_list_base.hpp"
 #include "../field/field_traits.hpp"
+// #include "field_list_metadata.hpp"
 #include "field_list_lookup.hpp"
 #include "../lib/metaprog/mp.hpp"
